add portaudio/19.7.0 + modernize (#1400)
* Move 19.6.0 to own folder

* Recreate all folder

* Start modernizing recipe

* Use CMake also for Linux and macOS builds

* Exclude Apple-Clang < 11

* Delete FindPortaudio.cmake

* Remove cpp_bindings option for now

As there is no support in CMake and according to the wiki:

"You do not need to use any special bindings to use PortAudio in C++ code."

https://github.com/PortAudio/portaudio/wiki/Tips_CPlusPlus

* Exclude apple-clang < 11 for 19.6.0 too

I'm really not motivated to track down what change in the environment is incompatible with Portaudio for compiler versions potentially nobody cares about.

Contributions are always welcome.

* Don't delete CMake + pkgconfig files for now

* Actual copy license file

* Rename folder all -> 19.7.0 to prepare for 19.8.0